requirements.txt
flask
opencv-python
dlib
torch
torchvision
pillow
numpy
tqdm